requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
